Split Projectile.cs into separate files and fix nullability warnings
